Introduce asr1k configmap to be read from all agents and server
Some settings need to be read from the neutron-server as well as from
the asr1k agents. To make a clean cut we introduce a new configmap for
that. We also introduce the flag `ignore_invalid_az_hint_for_router`
which will tell the driver to still schedule a router even if the
requested AZ has no asr1k agent available. This is a transition feature
so the API can be offered but the technical necessities in the backend
are not present yet.